# socialdrinks-cocktails/cocktails-web/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Socialdrinks Cocktails API",
    "version": "1.0.0",
    "description": "REST-API für Cocktails, Zutaten und den persönlichen Kühlschrank. Die Definition ist statisch gepflegt und enthält zusätzliche Beispiele für Antworten, die sich aus der Implementierung nicht besonders gut automatisch ableiten lassen."
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "Lokaler WildFly"
    }
  ],
...
  "paths": {
...
    "/api/cocktails/{id}": {
      "get": {
        "tags": [
          "Cocktails"
        ],
        "summary": "Cocktail per ID abrufen",
        "description": "Liefert einen Cocktail inklusive seiner Instruktionen und Zutaten.",
        "operationId": "getCocktail",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "ID des Cocktails",
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Cocktail gefunden",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CocktailDetails"
                },
                "examples": {
                  "pinkPower": {
                    "value": {
                      "id": 26,
                      "name": "Pink Power",
                      "instructions": [
                        {
                          "amountCL": 4,
                          "ingredient": {
                            "id": 7,
                            "name": "Fruchtsirup Preiselbeere"
                          }
                        },
                        {
                          "amountCL": 15,
                          "ingredient": {
                            "id": 30,
                            "name": "Buttermilch"
                          }
                        }
                      ]
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/cocktails/search": {
      "get": {
        "tags": [
          "Cocktails"
        ],
        "summary": "Cocktails suchen",
        "description": "Durchsucht Cocktails und Zutatennamen. Als Suchbegriff eignet sich in der Praxis ein Text ab etwa drei Zeichen.",
        "operationId": "searchCocktails",
        "parameters": [
          {
            "name": "query",
            "in": "query",
            "required": true,
            "description": "Suchbegriff",
            "schema": {
              "type": "string"
            },
            "example": "Pink"
          }
        ],
        "responses": {
          "200": {
            "description": "Suchtreffer",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/CocktailSummary"
                  }
                },
                "examples": {
                  "default": {
                    "value": [
                      {
                        "id": 26,
                        "name": "Pink Power"
                      }
                    ]
                  }
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "CocktailSummary": {
        "type": "object",
        "required": [
          "id",
          "name"
        ],
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64",
            "example": 26
          },
          "name": {
            "type": "string",
            "example": "Pink Power"
          }
        }
      },
      "Ingredient": {
        "type": "object",
        "required": [
          "id",
          "name"
        ],
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64",
            "example": 30
          },
          "name": {
            "type": "string",
            "example": "Buttermilch"
          }
        }
      },
      "Instruction": {
        "type": "object",
        "properties": {
          "amountCL": {
            "type": "integer",
            "example": 15
          },
          "ingredient": {
            "$ref": "#/components/schemas/Ingredient"
          }
        }
      },
      "CocktailDetails": {
        "allOf": [
          {
            "$ref": "#/components/schemas/CocktailSummary"
          },
          {
            "type": "object",
            "properties": {
              "instructions": {
                "type": "array",
                "items": {
                  "$ref": "#/components/schemas/Instruction"
                }
              }
            }
          }
        ]
      },
...
    }
  }
}